requirements.txt
absl-py==0.10.0
aniso8601==8.0.0
astunparse==1.6.3
autopep8==1.5.4
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
flake8==3.8.3
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
gast==0.3.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
IMDbPY==6.8
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.4.3
Keras-Preprocessing==1.1.2
line-bot-sdk==1.16.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.1
Pillow==7.2.0
pip-autoremove==0.9.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
SQLAlchemy==1.3.19
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-cpu==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
